Resolve circular reference by moving consistency code to path_data
